+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/report/ReportUtils.groovy
@@ -47,6 +47,8 @@ class ReportUtils {
      * @return file link to use in console output
      */
     static String toConsoleLink(File file) {
-        return "file:///${file.canonicalPath.replaceAll('\\\\', '/')}"
+        String path = file.canonicalPath.replaceAll('\\\\', '/')
+        // remove trailing slash for linux
+        return "file:///${path.startsWith('/') ? path[1..-1] : path}"
     }
 }